Add Functor instances for Coyoneda, {Co}PastroSum, Environment, {Co}Pastro, and Free{Mapping,Traversing} (#93)
These are quite straightforward to define, and they will likely be useful
in a future where `Profunctor` has a quantified `Functor` superclass, such
as what is being proposed in #70. In fact, this patch is simply a subset
of that one, authored by ChShersh.

There are other data types that could potentially have `Functor` instances,
but the exact shapes of these instances is yet unclear. (See the ongoing
discussion in #70.) For now, I've opted to stick with the most uncontroversial
instances that do not use any flexible or quantified constraints.

This supersedes #92, which requests the `Coyoneda` instance in particular.
To ensure that these instances are forward compatible with a proposal like #70,
I have rewritten the proposed instance in #92 to avoid the use of
`fmap = rmap`, as it is quite likely that #70 will introduce a new `Profunctor`
law that `rmap = fmap`.

Co-authored-by: ChShersh <dmitrii@holmusk.com>
Co-authored-by: Ole Krüger <ole@vprsm.de>